commit home to github

+++ src/com/kuteinykov/programm/services/ContactService.java
@@ -4,8 +4,8 @@ public interface ContactService {
 
     void initContact();
     void createContact(String name, String phoneNumber, int age, String address);
-    void deleteContact(int id);
+    void deleteContact(long id);
     void displayContact();
-    void editContact(int id, String name, String phoneNumber,int age, String address);
+    void editContact(long id, String name, String phoneNumber, int age, String address);
     void findContact(String name);
 }
